Central agent work-load partitioning
Provides a mechanism to allow the central agent to be horizontally
scaled out, such that each agent polls a disjoint subset of resources.

This is achieved through the use of `tooz` library for distributed
coordination.

If a service wants to use work-load partitioning, it must first
create a PartitionCoordinator object and set it up with the
`set_partition_coordinator` method of the `partitioning` module.
It must also call its `heartbeat` method periodically.

To distribute a set of resources over multiple agents, decorate
a function returning that set with the Partition decorator, for
example:

 @Partition
 def discovery():
     return [...]

The decorator uses the global `PartitionCoordinator` to figure out
which agents are in the same group and figures out which resources
belong to the current agent.

Change-Id: I7adef87b03129f4f8b38109bf547c7403cc6adad
Implements: blueprint central-agent-partitioning